% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/as.DataTables-method.R
\docType{methods}
\name{as.DataTables}
\alias{as.DataTables}
\alias{as.DataTables,data.frame-method}
\title{turn object into html using DataTables}
\usage{
as.DataTables(object, ...)

\S4method{as.DataTables}{data.frame}(object, align = NULL, jsDir = NULL)
}
\arguments{
\item{object}{the relevant object}

\item{...}{further arguments}

\item{align}{defaults to null, if supplied, same length as table}

\item{jsDir}{pass a specified path}
}
\value{
a html object
}
\description{
a story to be told
}

