% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{browse}
\alias{browse}
\alias{browse,data.frame-method}
\alias{browse,html-method}
\title{view html object in a browser}
\usage{
browse(object, ...)

\S4method{browse}{html}(object)

\S4method{browse}{data.frame}(object)
}
\arguments{
\item{object}{the object to be viewed}

\item{...}{further arguments}
}
\description{
a story to be told
}

