% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/DataTablesR-package.R
\docType{package}
\name{DataTablesR-package}
\alias{DataTablesR}
\alias{DataTablesR-package}
\title{DataTablesR-package}
\description{
R wrapper for DataTables
}
\author{
Andreas Blaette (andreas.blaette@uni-due.de)
}
\references{
http://polmine.sowi.uni-due.de
}
\keyword{package}

