% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{as.DataTables}
\alias{as.DataTables}
\alias{as.DataTables,data.frame-method}
\title{turn object into html using DataTables}
\usage{
as.DataTables(object, ...)

\S4method{as.DataTables}{data.frame}(object)
}
\arguments{
\item{object}{the relevant object}

\item{...}{further arguments}
}
\value{
a html object
}
\description{
a story to be told
}

